Add "taken_time" and associated timestamp handling functions
This commit introduces the "taken_time" field in the data.json file to record the time when a person was taken. Two functions, parseCustomDateFormat() and getRelativeTime(), were added in scripts.js to deal with this new timestamp. Both the JSON data structure and the README file were well updated to reflect these changes.